telegram-notifier/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
aiogram==3.13.1
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3